Refactor Classes and Add Packages
