+++ src/main/duke/command/Command.java
@@ -1,0 +1,5 @@
+package duke.command;
+
+public class Command {
+    public static
+}
