+++ src/main/duke/parser/Parser.java
@@ -1,0 +1,6 @@
+public class Parser {
+
+    public static String[] commandToArray(String text){
+        return text.split(" ");
+    }
+}
